Change the folder structure
