+++ examples/mimc.go
@@ -2,6 +2,7 @@ package examples
 
 import (
 	"github.com/consensys/gkr-mimc/circuit"
+	"github.com/consensys/gkr-mimc/circuit/gates"
 	"github.com/consensys/gkr-mimc/hash"
 )
 
@@ -12,14 +13,14 @@ func CreateMimcCircuit() circuit.Circuit {
 
 	for i := 0; i < nRounds-1; i++ {
 		wiring[i] = []circuit.Wire{
-			circuit.Wire{L: 1, R: 0, O: 0, Gate: circuit.NewCipherGate(hash.Arks[i])},
-			circuit.Wire{L: 1, R: 0, O: 1, Gate: circuit.CopyGate{}},
+			{L: 1, R: 0, O: 0, Gate: gates.NewCipherGate(hash.Arks[i])},
+			{L: 1, R: 0, O: 1, Gate: gates.CopyGate{}},
 		}
 	}
 
 	// And we don't copy the input in the last layer
 	wiring[nRounds-1] = []circuit.Wire{
-		circuit.Wire{L: 1, R: 0, O: 0, Gate: circuit.NewCipherGate(hash.Arks[nRounds-1])},
+		{L: 1, R: 0, O: 0, Gate: gates.NewCipherGate(hash.Arks[nRounds-1])},
 	}
 
 	return circuit.NewCircuit(wiring)
